fix: readme, move functions in util.hpp to the library, add tips in yaml file
