+++ bbs3d/include/pointcloud_iof/pcd_loader_without_pcl.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include <pointcloud_iof/io.hpp>
+#include <pointcloud_iof/pcd_io.hpp>
 #include <pointcloud_iof/filter.hpp>
 #include <boost/filesystem.hpp>
