--- ros2_test/rviz2/include/util.hpp
@@ -1,61 +1,0 @@
-#pragma once
-#include <pcl/filters/voxel_grid.h>
-#include <pcl/filters/approximate_voxel_grid.h>
-#include <pcl/io/pcd_io.h>
-#include <pcl/common/transforms.h>
-
-#include <tf2_ros/transform_broadcaster.h>
-#include <geometry_msgs/msg/transform_stamped.hpp>
-
-std::string createDate() {
-  time_t t = time(nullptr);
-  const tm* local_time = localtime(&t);
-  std::stringstream s;
-  s << "20" << local_time->tm_year - 100 << "_";
-  s << local_time->tm_mon + 1 << "_";
-  s << local_time->tm_mday << "_";
-  s << local_time->tm_hour << "_";
-  s << local_time->tm_min << "_";
-  s << local_time->tm_sec;
-  return (s.str());
-}
-
-bool load_tar_clouds(const std::string& tar_path, const float tar_leaf_size, pcl::PointCloud<pcl::PointXYZ>::Ptr& tar_cloud_ptr) {
-  // Load pcd file
-  boost::filesystem::path dir(tar_path);
-  if (!boost::filesystem::exists(dir)) {
-    std::cout << "[ERROR] Can not open folder" << std::endl;
-    return false;
-  }
-
-  pcl::PointCloud<pcl::PointXYZ>::Ptr cloud_ptr(new pcl::PointCloud<pcl::PointXYZ>());
-  for (const auto& file : boost::filesystem::directory_iterator(tar_path)) {
-    const std::string filename = file.path().c_str();
-    const std::string extension = file.path().extension().string();
-    if (extension != ".pcd" && extension != ".PCD") {
-      continue;
-    }
-
-    // Check load pcd
-    pcl::PointCloud<pcl::PointXYZ>::Ptr cloud_temp_ptr(new pcl::PointCloud<pcl::PointXYZ>());
-    if (pcl::io::loadPCDFile(filename, *cloud_temp_ptr) == -1) {
-      std::cout << "[WARN] Can not open pcd file: " << filename << std::endl;
-      continue;
-    }
-    *cloud_ptr += *cloud_temp_ptr;
-  }
-
-  // Downsample
-  if (tar_leaf_size != 0.0f) {
-    pcl::PointCloud<pcl::PointXYZ>::Ptr filtered_cloud_ptr(new pcl::PointCloud<pcl::PointXYZ>());
-    pcl::ApproximateVoxelGrid<pcl::PointXYZ> filter;
-    filter.setLeafSize(tar_leaf_size, tar_leaf_size, tar_leaf_size);
-    filter.setInputCloud(cloud_ptr);
-    filter.filter(*filtered_cloud_ptr);
-    *tar_cloud_ptr = *filtered_cloud_ptr;
-  } else {
-    *tar_cloud_ptr = *cloud_ptr;
-  }
-
-  return true;
-}
