+++ test/src/gpu_test.cpp
@@ -1,8 +1,8 @@
 #include <gpu_bbs3d/bbs3d.cuh>
 #include <pointcloud_iof/pcl_eigen_coverter.hpp>
+#include <pointcloud_iof/pcd_loader.hpp>
 
 #include <test.hpp>
-#include <util.hpp>
 #include <load_yaml.hpp>
 #include <chrono>
 
@@ -10,6 +10,19 @@
 #include <pcl/io/pcd_io.h>
 #include <pcl/point_types.h>
 #include <pcl/registration/gicp.h>
+
+std::string create_date() {
+  time_t t = time(nullptr);
+  const tm* local_time = localtime(&t);
+  std::stringstream s;
+  s << "20" << local_time->tm_year - 100 << "_";
+  s << local_time->tm_mon + 1 << "_";
+  s << local_time->tm_mday << "_";
+  s << local_time->tm_hour << "_";
+  s << local_time->tm_min << "_";
+  s << local_time->tm_sec;
+  return (s.str());
+}
 
 BBS3DTest::BBS3DTest() {}
 
@@ -25,7 +38,7 @@ int BBS3DTest::run(std::string config) {
   // Load target pcds
   std::cout << "[Setting] Loading target pcds..." << std::endl;
   pcl::PointCloud<pcl::PointXYZ>::Ptr tar_cloud_ptr(new pcl::PointCloud<pcl::PointXYZ>());
-  if (!load_tar_clouds(tar_path, tar_leaf_size, tar_cloud_ptr)) {
+  if (!pciof::load_tar_clouds(tar_path, tar_leaf_size, tar_cloud_ptr)) {
     std::cout << "[ERROR] Loading target pcd failed" << std::endl;
     return 1;
   }
@@ -43,14 +56,14 @@ int BBS3DTest::run(std::string config) {
   // Load source pcds
   std::cout << "[Setting] Loading source pcds..." << std::endl;
   std::vector<std::pair<std::string, pcl::PointCloud<pcl::PointXYZ>::Ptr>> src_cloud_set;
-  if (!load_src_clouds(src_path, min_scan_range, max_scan_range, src_leaf_size, src_cloud_set)) {
+  if (!pciof::load_src_points_with_filename(src_path, min_scan_range, max_scan_range, src_leaf_size, src_cloud_set)) {
     std::cout << "[ERROR] Loading source pcds failed" << std::endl;
     return 1;
   };
 
   // Create output folder with date
   std::cout << "[Setting] Create output folder with date..." << std::endl;
-  std::string date = createDate();
+  std::string date = create_date();
   std::string pcd_save_folder_path = output_path + "/" + date;
   boost::filesystem::create_directory(pcd_save_folder_path);
