+++ test/test_code.md
@@ -54,8 +54,8 @@ You can save the voxelmaps coordinates and skip 3D-BBS voxel construction if you
 cd 3d_bbs/test/build/
 ./saver <target folder path> min_level_res max_level 
 ```
-Then, try again 4.Run.  
-Note: Give priority to parameters in the generated text file.
+Then, try again 4.Run. Voxelmaps coordinates in 'voxelmaps_coords' are automatically loaded.  
+Note: Test code preferentially uses the parameters in the generated text file 'voxel_params.txt'.
 
 ## Using your own 3D point cloud map and LiDAR scan data
 ### Convert ros2 bag to pcd file
@@ -69,6 +69,3 @@ https://github.com/KOKIAOKI/ros2bag_to_pcd
 - Use only at locations where the source point cloud is completely included in the target point cloud.
   - Reason: Another pose that encompasses all source point cloud is estimated when the source point cloud includes outside the map environment. 
   - Please use the downsampling and point cloud cutting tools.
-
-  ### Create voxelmaps coord in advance
-
